+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # dev:
 - only use hydra on main function of ip_one_tile / add_buffer_one_tile / filter_one_tile
+- add possibility to overwrite no-data in polygons from a shapefile after interpolation (eg. for ZICAD)
 
 # v0.2.3:
 - Continuous Integration with Jenkins